backend/reset-admin/requirements.txt
psycopg2-binary>=2.9.0
bcrypt
